sql_query_app/requirements.txt
streamlit
pyodbc
pandas
openpyxl
bcrypt
python-dotenv
